Increase the memory of SORMADUP with the number of CPUs
Contrary to my first, limited, assessment of the SORMADUP composite module,
the memory usage does increase with the number of CPUs.

We found that after 1 SORMADUP failed because of MEMLIMIT at every attempt.

I reran the job with all CPU requirements and checked the memory usage

```
 8  22,429 MB
14  27,303 MB
20  32,383 MB
26  37,089 MB
32  41,990 MB
38  46,931 MB
```
The correlation line is y=816*x+15929

I rounded it up to 850 MB per CPU and decreased the base offset accordingly
so that the base memory assignment remains the same (after all, it's only
1 failure out of many species).
However, I made sure that the per-CPU factor increases every attempt, so that
that species would have succeeded at the second attempt.