More memory for crumble !
The previous code was attempting Crumble with 1 GB, 3, 6, 10, 15, 21 (with the
recent addition of a multiplier based on the number of reads).

Across the recent runs, a bunch of species failed at 21 GB. They needed 29 GB.
Other species failed at 42 GB. They needed 48, 51, 58, 59, or 108 GB.

The formula 3^(attempt-1) will provide the following sequence:
 1 GB, 3, 9, 27, 81, 243
It provides a >2-fold cushion at the high end of the distribution.
If it still fails, we would then move to 4^(attempt-1):
 1 GB, 4, 16, 64, 256, 1024.
which I want to avoid for now in order to minimise wastage.